Create Convert Byte Array to Hexadecimal
In the above program, we have a byte array named bytes. To convert byte array to a hex value, we loop through each byte in the array and use String's format().

We use %02X to print two places (02) of Hexadecimal (X) value and store it in the string st.

This is a relatively slower process for large byte array conversion. We can dramatically increase the speed of execution using byte operations shown below.